Implicit cast

+++ main.c
@@ -232,14 +232,14 @@ dowork(char *file, size_t nfile)
 	struct station *result = merge(g_stations[0], nthread-1);
 	qsort(result, MAX_CAPACITY, sizeof *result, compare);
 	double avg = (double)result[0].sum / result[0].cnt * 0.1;
-	double min = (double)result[0].min * 0.1;
-	double max = (double)result[0].max * 0.1;
+	double min = result[0].min * 0.1;
+	double max = result[0].max * 0.1;
 	printf("{%.*s=%.1f/%.1f/%.1f", result[0].nname, result[0].name, min, avg, max);
 	for (size_t i = 1; i < MAX_CAPACITY; i++) {
 		if (!result[i].cnt) break;
 		avg = (double)result[i].sum / result[i].cnt * 0.1;
-		min = (double)result[i].min * 0.1;
-		max = (double)result[i].max * 0.1;
+		min = result[i].min * 0.1;
+		max = result[i].max * 0.1;
 		printf(", %.*s=%.1f/%.1f/%.1f", result[i].nname, result[i].name, min, avg, max);
 	}
 	printf("}\n");
